moving database interactions outside the database connections class
